sql/catalog/catalogkv: move UncachedPhysicalAccessor to catalogkv
It's an interface to access the database.

Release note: None
